+++ config/container.php
@@ -57,7 +57,7 @@ return [
         $translator = $container->get(Translator::class);
         $twig->addExtension(new TranslationExtension($translator));
 
-        $twig->addExtension(new TwigAssetVersionExtension(__DIR__ . '/../public/assets/rev-manifest.json'));
+        $twig->addExtension(new TwigAssetVersionExtension(__DIR__ . '/../public/assets/mix-manifest.json'));
 
         $twig->offsetSet('domain', $_SERVER['HTTP_HOST']);
         $twig->offsetSet('host', $_SERVER['REQUEST_SCHEME'] . '://' . $_SERVER['HTTP_HOST']);