+++ config/container.php
@@ -66,6 +66,7 @@ return [
             'denominations' => Configuration::getDenominations(),
             'types' => Configuration::getTypes(),
             'websiteTypes' => Configuration::getWebsiteTypes(),
+            'websiteTypeIcons' => Configuration::getWebsiteTypeIcons(),
             'startOfWebsiteURL' => Configuration::getStartOfWebsiteURL(),
             'websitesTypesToCompare' => Configuration::getWebsiteTypesToCompare(),
             'websitesToCompareColors' => Configuration::getWebsitesToCompareColors(),