+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Update dependencies
+* Switch from Gulp to Laravel Mix as build tool
 
 
 ## Version 3.7.2 (2021-05-27)