+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Replace jQuery based tablesorter with tablesort library (which has no dependencies)
 - Remove jQuery
 - Add icons to menu items, replace icons for social networks and other website types
+- Add website type "podcast"
 
 
 ## Version 3.7.2 (2021-05-27)