+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switch from Gulp to Laravel Mix as build tool
 - Replace jQuery based tablesorter with tablesort library (which has no dependencies)
 - Remove jQuery
+- Improve layout of details page
+- Translate month names into German (in the follower history chart)
 
 
 ## Version 3.7.2 (2021-05-27)